tools/requirements.txt
anyio==4.3.0
certifi==2024.7.4
click==8.1.7
fake-useragent==1.5.1
geographiclib==2.0
geojson==3.1.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-magic==0.4.27
shapely==2.0.4
sniffio==1.3.1
